improved the input password
