+++ src/Components/Login/SignupBlock.jsx
@@ -15,7 +15,6 @@ const SignupBlock = ({ setIsSignUp }) => {
     password: "",
   });
   const [isValidationError, setIsValidationError] = useState(false);
-  const [showPassword, setShowPassword] = useState(false);
 
   const {
     axiosInstance,
@@ -80,10 +79,6 @@ const SignupBlock = ({ setIsSignUp }) => {
     setIsSignUp(false);
   };
 
-  const togglePasswordVisibility = () => {
-    setShowPassword(!showPassword);
-  };
-
   return (
     <>
       <h1 className="text-lg font-bold">Create your account</h1>
@@ -102,24 +97,12 @@ const SignupBlock = ({ setIsSignUp }) => {
         error="Email required"
       />
 
-      <div className="relative">
-        <Input
-          label="Password"
-          type={showPassword ? "text" : "password"}
-          onChange={(e) => onChangeHandler("password", e.target.value)}
-          showError={isValidationError && !regDetails.password}
-          error="Password Required"
-        />
-        <button
-          type="button"
-          onClick={togglePasswordVisibility}
-          className="absolute right-2 top-7 text-sm text-violet-600 focus:outline-none"
-        >
-          {showPassword ? "Hide" : "Show"}
-        </button>
-      </div>
-
-
+      <Input
+        label="Password"
+        onChange={(e) => onChangeHandler("password", e.target.value)}
+        showError={isValidationError && !regDetails.password}
+        error="Password Required"
+      />
 
       <Button isLoading={isLoading} onClick={signUpBtnHandler} name="SIGN UP" />
